Voice Assistant - V1
This is version 1 of the voice assistant. It uses the speech recognition module to pick up audio from the microphone  and then uses the recogniser within the module to recognise the audio, repeat it in the console and speak it using the speaktext() subroutine. After It would check for exception errors based on the problems in the code.