Voice Assistant V2
Here the program has not changed much in terms of recognising and listening for audio from the microphone. The change has come in how the audio is output. The assistantVoice() subroutine uses the google text to speech module gTTS to create an audio file with the desired output. The audio is then saved to a file, run using the OS module and then deleted. Currently an inefficient solution to the problem however it shall be developed in coming stages.