Voice Assistant V3
In this epoch of the voice assistant the audio output system has once again been changed. Instead of the inefficient method of constantly creating and deleting audio files containing the output audio, the pyttsx3 module has been used to make quick work of the issue. By simply creating an instance of the class and using it's say() method we are able to cut down time to run substantially and remove many possible points of crashes and error. Both the audio output and the conversion of audio to text have also been put into subroutines. A few small subroutines have also been added to test the implementation of requests to the voice assistant