Voice Assistant V3.2
Here big improvements have been made to the system. More subroutines have been added to give the voice assistant more functionality and allow the assistant to return the day and date which it is currently (This has been done using the datetime module). A new subroutine is now being used to process requests called "process_query()". This calls the getAudio() subroutine and converts what it returns to a string. From there it analyses it for words which correspond to actions it can perform and calls the appropriate subroutine. These subroutines form the output and send it to the assistantVoice function to be played.