Voice Assistant V3.3
Here more functionality has been added with the addition of the settings file. This allows the user to change the time format which is used in the system to their liking as they use the application. There is also now the return_time() subroutine which will return the specific time in their desired format (currently unfinished but that is the intention). Also not enough comments but they are coming along