C.A.R.I Control Centre 1.0.2
This has been updated massively from the initial version. The grid and layout of the application has now been set up and pages have been implemented in the form of subroutines. The creation of buttons has been implemented in the ui_creation.py file instead for simplicity on my end in terms of the amount of lines of code that would be in the main app page. As well as all this a colour scheme has been established to give the app a more professional and well thought out look.